UBX + RTCM improvements
   * introduce rtcm crate feature, to attempt RTCM decoding
   * add more comments to existing parsers and structures,
   in particular that UbxRtcm handles both and will adapt on the fly
